+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.0] - (06.10.2024)
+
+## Changes
+
+- add buttons, sensors and presets to Mitsubishi AC device (MiElHVAC)
+- config.schema update
+- redme update
+- cleanup
+
 ## [0.14.0] - (05.10.2024)
 
 ## Changes